airflow/requirements.txt
pandas
apache-airflow==2.7.0
scikit-learn==1.4.0
tensorflow-cpu==2.13.0
keras==2.13.1
mlflow==2.4.1
joblib
psycopg2==2.9.10
psycopg2-binary==2.9.9
SQLAlchemy
dill==0.3.8
opencv-python
pillow==10.0.0